backend/pyproject.toml
[tool.poetry]
name = "CU-CRASH"
version = "0.1.0"
description = "backend using python fastapi for cu-crash project in soft arch 2021"
authors = ["Jobjobsss <chayapoi-p@hotmail.com>"]
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {extras = ["all"], version = "^0.70.0"}
SQLAlchemy = "^1.4.25"
mysqlclient = "^2.0.3"
grpcio = "^1.40.0"
grpcio-tools = "^1.40.0"
PyJWT = "^2.1.0"
bcrypt = "^3.2.0"
pyodbc = "^4.0.32"
pymongo = "^3.12.1"
pydantic = {extras = ["dotenv"], version = "^1.8.2"}
email-validator = "^1.1.3"
python-multipart = "^0.0.5"
uvicorn = "^0.15.0"
pika = "^1.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "app.main:start"
start_grpc = "app.scripts:start_grpc_server"
